removed "lowMem3" and made "copy" into main document
